requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==4.2.13
django-appconf==1.0.6
django-compressor==4.5.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
django-summernote==0.8.20.0
mysqlclient==2.2.7
pillow==10.3.0
python-decouple==3.8
rcssmin==1.1.2
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
webencodings==0.5.1
